refactor: move common test utils from cli to testing-utils
